Move requestAnimation in its own file for better reuse
